task2/metadata/requirements.txt
PyYAML==6.0.2
pandas==2.0.1